requirements.txt
asgiref==3.6.0
attrs==23.1.0
black==23.3.0
certifi==2022.9.24
charset-normalizer==3.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
Django==4.2
django-filter==23.1
django-mail-templated==2.6.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
filelock==3.8.0
idna==3.4
importlib-metadata==6.6.0
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
Markdown==3.4.3
MarkupSafe==2.1.2
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
pipenv==2022.11.30
platformdirs==2.5.4
PyJWT==2.7.0
pyrsistent==0.19.3
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
requests==2.30.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
swagger-spec-validator==3.0.3
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
virtualenv==20.17.0
virtualenv-clone==0.5.7
zipp==3.15.0
